Fix: Behat tests fail with signature validation errors
Since simplesamlphp/saml2 3.2.0 RSA_SHA256 is the default, this somehow breaks verification with a key that is loaded with RSA_SHA1, where this worked before.
Because all signatures should in fact be signed using RSA SHA-265, we should use that. This works with saml2 version before 3.2.0 as well.
Tested with surfnet/stepup-saml-bundle 4.1.2